requirements.txt
Django==5.0.4
psycopg2-binary==2.9.10
Pillow==11.0.0
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
PyJWT==2.11.0
